GH-2: Rearranged things to allow previous `node .` syntax.
