+++ README.md
@@ -12,7 +12,7 @@ Before you can use this package, you will need to install its dependencies using
 
 To launch the router, use a command like:
 
-`node launcher.js`
+`node .`
 
 If you launch the command without any arguments (or with the `--help` option), usage information will be displayed.
 
@@ -22,7 +22,7 @@ If you launch the command without any arguments (or with the `--help` option), u
 To use this command, you will need to know part or all of the name of the attached input and output devices.  You can
 display all detected inputs and outputs using a commmand like:
 
-`node launcher.js -l`
+`node . -l`
 
 A list of outputs like the following will be displayed:
 
@@ -62,12 +62,12 @@ To route ports, pass in one or more inputs, and one or more outputs, as in the f
 the above example:
 
 ```
-node launcher.js -i Oxygen -o "MIDI Monitor"
+node . -i Oxygen -o "MIDI Monitor"
 ```
 
 Note the quotation marks around the output.  You can also pass in space-delimited arrays, as in the following example
 (which assumes you have inputs named `input1` and `input 2`, and outputs named `output 1` and `output2`:
 
 ```
-node launcher.js -i input1 "input 2" -o "output 1" output2
+node . -i input1 "input 2" -o "output 1" output2
 ```
